# Improve execution speed of BBM stress update