--- a/dynamics/src/ParametricMap.cpp
+++ b/dynamics/src/ParametricMap.cpp
@@ -84,7 +84,7 @@ template <int DG> void ParametricTransportMap<DG>::InitializeInverseDGMassMatrix
 //////////////////////////////////////////////////
 
 //!
-  template <int CG, int DG> void ParametricMomentumMap<CG, DG>::InitializeLumpedCGMassMatrix()
+template <int CG, int DG> void ParametricMomentumMap<CG, DG>::InitializeLumpedCGMassMatrix()
 {
     lumpedcgmass.resize_by_mesh(smesh);
 
@@ -154,7 +154,7 @@ template <int DG> void ParametricTransportMap<DG>::InitializeInverseDGMassMatrix
 }
 
 //!
-  template <int CG, int DG> void ParametricMomentumMap<CG, DG>::InitializeDivSMatrices()
+template <int CG, int DG> void ParametricMomentumMap<CG, DG>::InitializeDivSMatrices()
 {
     divS1.resize(smesh.nelements);
     divS2.resize(smesh.nelements);
@@ -221,14 +221,14 @@ template <int DG> void ParametricTransportMap<DG>::InitializeInverseDGMassMatrix
                 * (PSI<CG2DGSTRESS(CG), GAUSSPOINTS1D(CG2DGSTRESS(CG))>.array().rowwise()
                     * (GAUSSWEIGHTS<GAUSSPOINTS1D(CG2DGSTRESS(CG))>.array() * J.array()))
                       .matrix();
-	    // same but for the damage. However, we use the same number of Gausspoints as
-	    // for the DG-stress variant above for easier use in BBM Stress update	   
-	    const Eigen::Matrix<Nextsim::FloatType, DG, DG> imass_dam
-	      = ParametricTools::massMatrix<DG>(smesh, eid).inverse();
-	    iMJwPSI_dam[eid] = imass_dam
-	      * (PSI<DG, GAUSSPOINTS1D(CG2DGSTRESS(CG))>.array().rowwise()
-		 * (GAUSSWEIGHTS<GAUSSPOINTS1D(CG2DGSTRESS(CG))>.array() * J.array()))
-	      .matrix();
+            // same but for the damage. However, we use the same number of Gausspoints as
+            // for the DG-stress variant above for easier use in BBM Stress update
+            const Eigen::Matrix<Nextsim::FloatType, DG, DG> imass_dam
+                = ParametricTools::massMatrix<DG>(smesh, eid).inverse();
+            iMJwPSI_dam[eid] = imass_dam
+                * (PSI<DG, GAUSSPOINTS1D(CG2DGSTRESS(CG))>.array().rowwise()
+                    * (GAUSSWEIGHTS<GAUSSPOINTS1D(CG2DGSTRESS(CG))>.array() * J.array()))
+                      .matrix();
 
         } else if (smesh.CoordinateSystem == SPHERICAL) {
             // In spherical coordinates (x,y) coordinates are (lon,lat) coordinates
@@ -271,15 +271,13 @@ template <int DG> void ParametricTransportMap<DG>::InitializeInverseDGMassMatrix
                     * (GAUSSWEIGHTS<GAUSSPOINTS1D(CG2DGSTRESS(CG))>.array() * J.array()))
                       .matrix();
 
-	    // smae for DG advection (damage)
-	    const Eigen::Matrix<Nextsim::FloatType, DG, DG> imass_dam
+            // smae for DG advection (damage)
+            const Eigen::Matrix<Nextsim::FloatType, DG, DG> imass_dam
                 = SphericalTools::massMatrix<DG>(smesh, eid).inverse();
             iMJwPSI_dam[eid] = imass_dam
                 * (PSI<DG, GAUSSPOINTS1D(CG2DGSTRESS(CG))>.array().rowwise()
                     * (GAUSSWEIGHTS<GAUSSPOINTS1D(CG2DGSTRESS(CG))>.array() * J.array()))
                       .matrix();
-
-
 
         } else
             abort();
@@ -291,11 +289,11 @@ template class ParametricTransportMap<3>;
 template class ParametricTransportMap<6>;
 template class ParametricTransportMap<8>;
 
-template class ParametricMomentumMap<1,1>;
-template class ParametricMomentumMap<2,1>;
-template class ParametricMomentumMap<1,3>;
-template class ParametricMomentumMap<2,3>;
-template class ParametricMomentumMap<1,6>;
-template class ParametricMomentumMap<2,6>;
+template class ParametricMomentumMap<1, 1>;
+template class ParametricMomentumMap<2, 1>;
+template class ParametricMomentumMap<1, 3>;
+template class ParametricMomentumMap<2, 3>;
+template class ParametricMomentumMap<1, 6>;
+template class ParametricMomentumMap<2, 6>;
 
 }