--- a/dynamics/src/cgParametricMomentum.cpp
+++ b/dynamics/src/cgParametricMomentum.cpp
@@ -1,6 +1,6 @@
 /*!
  * @file ParametricMomentum.cpp
- * @date 1 Mar 2022
+ * @date 04 Sep 2024
  * @author Thomas Richter <thomas.richter@ovgu.de>
  */