--- a/dynamics/src/cgParametricMomentum.cpp
+++ b/dynamics/src/cgParametricMomentum.cpp
@@ -1,6 +1,6 @@
 /*!
  * @file ParametricMomentum.cpp
- * @date 04 Sep 2024
+ * @date 24 Sep 2024
  * @author Thomas Richter <thomas.richter@ovgu.de>
  */