--- a/dynamics/src/include/BBM.hpp
+++ b/dynamics/src/include/BBM.hpp
@@ -44,14 +44,13 @@ namespace BBM {
      */
     template <int CG, int DGs, int DGa>
     void StressUpdateHighOrder(const MEBParameters& params,
-			       const ParametricMomentumMap<CG, DGa>& pmap,
-			       const ParametricMesh& smesh,
-        DGVector<DGs>& S11, DGVector<DGs>& S12, DGVector<DGs>& S22, const DGVector<DGs>& E11,
-        const DGVector<DGs>& E12, const DGVector<DGs>& E22, const DGVector<DGa>& H,
-        const DGVector<DGa>& A, DGVector<DGa>& D, const double dt_mom)
+        const ParametricMomentumMap<CG, DGa>& pmap, const ParametricMesh& smesh, DGVector<DGs>& S11,
+        DGVector<DGs>& S12, DGVector<DGs>& S22, const DGVector<DGs>& E11, const DGVector<DGs>& E12,
+        const DGVector<DGs>& E22, const DGVector<DGa>& H, const DGVector<DGa>& A, DGVector<DGa>& D,
+        const double dt_mom)
     {
 
-//#define NGP (DGs == 8 ? 3 : (DGs == 3 ? 2 : -1))
+// #define NGP (DGs == 8 ? 3 : (DGs == 3 ? 2 : -1))
 #define NGP 3
 
 //! Stress and Damage Update
@@ -59,53 +58,53 @@ namespace BBM {
         for (size_t i = 0; i < smesh.nelements; ++i) {
 
             //! Evaluate values in Gauss points (3 point Gauss rule in 2d => 9 points)
-            const Eigen::Matrix<double, 1, NGP* NGP> h_gauss
+            const Eigen::Matrix<double, 1, NGP * NGP> h_gauss
                 = (H.row(i) * PSI<DGa, NGP>).array().max(0.0).matrix();
-            const Eigen::Matrix<double, 1, NGP* NGP> a_gauss
+            const Eigen::Matrix<double, 1, NGP * NGP> a_gauss
                 = (A.row(i) * PSI<DGa, NGP>).array().max(0.0).min(1.0).matrix();
-            Eigen::Matrix<double, 1, NGP* NGP> d_gauss
+            Eigen::Matrix<double, 1, NGP * NGP> d_gauss
                 = (D.row(i) * PSI<DGa, NGP>).array().max(1e-12).min(1.0).matrix();
 
-            const Eigen::Matrix<double, 1, NGP* NGP> e11_gauss = E11.row(i) * PSI<DGs, NGP>;
-            const Eigen::Matrix<double, 1, NGP* NGP> e12_gauss = E12.row(i) * PSI<DGs, NGP>;
-            const Eigen::Matrix<double, 1, NGP* NGP> e22_gauss = E22.row(i) * PSI<DGs, NGP>;
+            const Eigen::Matrix<double, 1, NGP * NGP> e11_gauss = E11.row(i) * PSI<DGs, NGP>;
+            const Eigen::Matrix<double, 1, NGP * NGP> e12_gauss = E12.row(i) * PSI<DGs, NGP>;
+            const Eigen::Matrix<double, 1, NGP * NGP> e22_gauss = E22.row(i) * PSI<DGs, NGP>;
 
-            Eigen::Matrix<double, 1, NGP* NGP> s11_gauss = S11.row(i) * PSI<DGs, NGP>;
-            Eigen::Matrix<double, 1, NGP* NGP> s12_gauss = S12.row(i) * PSI<DGs, NGP>;
-            Eigen::Matrix<double, 1, NGP* NGP> s22_gauss = S22.row(i) * PSI<DGs, NGP>;
+            Eigen::Matrix<double, 1, NGP * NGP> s11_gauss = S11.row(i) * PSI<DGs, NGP>;
+            Eigen::Matrix<double, 1, NGP * NGP> s12_gauss = S12.row(i) * PSI<DGs, NGP>;
+            Eigen::Matrix<double, 1, NGP * NGP> s22_gauss = S22.row(i) * PSI<DGs, NGP>;
 
             //! Current normal stress for the evaluation of tildeP (Eqn. 1)
-            Eigen::Matrix<double, 1, NGP* NGP> sigma_n
+            Eigen::Matrix<double, 1, NGP * NGP> sigma_n
                 = 0.5 * (s11_gauss.array() + s22_gauss.array());
 
             //! exp(-C(1-A))
-            const Eigen::Matrix<double, 1, NGP* NGP> expC
+            const Eigen::Matrix<double, 1, NGP * NGP> expC
                 = (params.compaction_param * (1.0 - a_gauss.array())).exp().array();
 
             // Eqn. 25
-            const Eigen::Matrix<double, 1, NGP* NGP> powalphaexpC
+            const Eigen::Matrix<double, 1, NGP * NGP> powalphaexpC
                 = (d_gauss.array() * expC.array()).pow(params.exponent_relaxation_sigma - 1);
-            const Eigen::Matrix<double, 1, NGP* NGP> time_viscous
+            const Eigen::Matrix<double, 1, NGP * NGP> time_viscous
                 = params.undamaged_time_relaxation_sigma * powalphaexpC;
 
             //! BBM  Computing tildeP according to (Eqn. 7b and Eqn. 8)
             // (Eqn. 8)
-            const Eigen::Matrix<double, 1, NGP* NGP> Pmax = params.P0
+            const Eigen::Matrix<double, 1, NGP * NGP> Pmax = params.P0
                 * h_gauss.array().pow(params.exponent_compression_factor) * expC.array();
 
             // (Eqn. 7b) Prepare tildeP
             // tildeP must be capped at 1 to get an elastic response
             // (Eqn. 7b) Select case based on sigma_n
-            const Eigen::Matrix<double, 1, NGP* NGP> tildeP
+            const Eigen::Matrix<double, 1, NGP * NGP> tildeP
                 = (sigma_n.array() < 0.0)
                       .select((-Pmax.array() / sigma_n.array()).min(1.0).matrix(), 0.);
 
             // multiplicator
-            const Eigen::Matrix<double, 1, NGP* NGP> multiplicator
+            const Eigen::Matrix<double, 1, NGP * NGP> multiplicator
                 = time_viscous.array() / (time_viscous.array() + (1. - tildeP.array()) * dt_mom);
 
             //! Eqn. 9
-            const Eigen::Matrix<double, 1, NGP* NGP> elasticity
+            const Eigen::Matrix<double, 1, NGP * NGP> elasticity
                 = h_gauss.array() * params.young * d_gauss.array() * expC.array();
 
             // Eqn. 12: first factor on RHS
@@ -115,7 +114,7 @@ namespace BBM {
              * \ (K:e)12 /    1 - nu^2 \  0   0  1-nu / \ e12 /
              */
 
-            const Eigen::Matrix<double, 1, NGP* NGP> Dunit_factor
+            const Eigen::Matrix<double, 1, NGP * NGP> Dunit_factor
                 = dt_mom * elasticity.array() / (1. - (params.nu0 * params.nu0));
 
             s11_gauss.array()
@@ -130,7 +129,7 @@ namespace BBM {
             s12_gauss.array() *= multiplicator.array();
 
             sigma_n = 0.5 * (s11_gauss.array() + s22_gauss.array());
-            const Eigen::Matrix<double, 1, NGP* NGP> tau
+            const Eigen::Matrix<double, 1, NGP * NGP> tau
                 = (0.25 * (s11_gauss.array() - s22_gauss.array()).square()
                     + s12_gauss.array().square())
                       .sqrt();
@@ -138,15 +137,15 @@ namespace BBM {
             const double scale_coef = std::sqrt(0.1 / smesh.h(i));
 
             //! Eqn. 22
-            const Eigen::Matrix<double, 1, NGP* NGP> cohesion
+            const Eigen::Matrix<double, 1, NGP * NGP> cohesion
                 = params.C_lab * scale_coef * h_gauss.array();
             //! Eqn. 30
-            const Eigen::Matrix<double, 1, NGP* NGP> compr_strength
+            const Eigen::Matrix<double, 1, NGP * NGP> compr_strength
                 = params.compr_strength * scale_coef * h_gauss.array();
 
             // Mohr-Coulomb failure using Mssrs. Plante & Tremblay's formulation
             // sigma_s + tan_phi*sigma_n < 0 is always inside, but gives dcrit < 0
-            Eigen::Matrix<double, 1, NGP* NGP> dcrit
+            Eigen::Matrix<double, 1, NGP * NGP> dcrit
                 = (tau.array() + params.tan_phi * sigma_n.array() > 0.)
                       .select(
                           cohesion.array() / (tau.array() + params.tan_phi * sigma_n.array()), 1.);
@@ -159,7 +158,7 @@ namespace BBM {
             dcrit = dcrit.array().min(1.0);
 
             // Eqn. 29
-            const Eigen::Matrix<double, 1, NGP* NGP> td = smesh.h(i)
+            const Eigen::Matrix<double, 1, NGP * NGP> td = smesh.h(i)
                 * std::sqrt(2. * (1. + params.nu0) * params.rho_ice) / elasticity.array().sqrt();
 
             // Update damage
@@ -171,7 +170,7 @@ namespace BBM {
             s22_gauss.array() -= s22_gauss.array() * (1. - dcrit.array()) * dt_mom / td.array();
 
             // // INTEGRATION OF STRESS AND DAMAGE
-            const Eigen::Matrix<Nextsim::FloatType, 1, NGP* NGP> J
+            const Eigen::Matrix<Nextsim::FloatType, 1, NGP * NGP> J
                 = ParametricTools::J<3>(smesh, i);
             // // get the inverse of the mass matrix scaled with the test-functions in the gauss
             // // points, with the gauss weights and with J. This is a 8 x 9 matrix
@@ -188,7 +187,7 @@ namespace BBM {
             S12.row(i) = pmap.iMJwPSI[i] * s12_gauss.matrix().transpose();
             S22.row(i) = pmap.iMJwPSI[i] * s22_gauss.matrix().transpose();
 
-            const Eigen::Matrix<Nextsim::FloatType, DGa, NGP* NGP> imass_psi2
+            const Eigen::Matrix<Nextsim::FloatType, DGa, NGP * NGP> imass_psi2
                 = ParametricTools::massMatrix<DGa>(smesh, i).inverse()
                 * (PSI<DGa, NGP>.array().rowwise() * (GAUSSWEIGHTS<NGP>.array() * J.array()))
                       .matrix();