--- a/dynamics/src/include/MEVPStressUpdateStep.hpp
+++ b/dynamics/src/include/MEVPStressUpdateStep.hpp
@@ -1,7 +1,7 @@
 /*!
  * @file MEVPStressUpdateStep.hpp
  *
- * @date 04 Sep 2024
+ * @date 24 Sep 2024
  * @author Tim Spain <timothy.spain@nersc.no>
  */