--- a/dynamics/src/include/ParametricMap.hpp
+++ b/dynamics/src/include/ParametricMap.hpp
@@ -1,6 +1,6 @@
 /*!
  * @file    ParametricMap.hpp
- * @date    Dec 13, 2022
+ * @date 04 Sep 2024
  * @author  Thomas Richter <thomas.richter@ovgu.de>
  */